Add time tracking functionality to TaskConsumer and TaskManager

+++ homework3/TaskManager.h
@@ -9,5 +9,5 @@ typedef struct TaskManager *TaskManagerPtr;
 TaskManagerPtr createTaskManager(int nLists);
 void destroyTaskManager(TaskManagerPtr taskManager);
 
-int taskManagerEstimateWaitingTime(TaskManagerPtr taskManager, int executionTime, size_t *pIndex);
+double taskManagerEstimateWaitingTime(TaskManagerPtr taskManager, int executionTime, size_t *pIndex);
 void taskManagerAddTask(TaskManagerPtr taskManager, int executionTime, sem_t *pCompletionSignal);
